requirements/base.txt
Django==3.0.5
django-environ==0.4.5
psycopg2-binary==2.8.4
django-braces==1.14.0
pypugjs==5.9.1
django-crispy-forms==1.9.0
django-datatables-view==1.19.1
